% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cism_map.R
\name{cism_map}
\alias{cism_map}
\title{cism map}
\usage{
cism_map(lng, lat, x = NULL, fspdf = NULL, type = NULL,
  make_simple = TRUE, n_simple = 10, opacity = 0.5, point_size = 1)
}
\arguments{
\item{lng}{A numeric vector of longitude coordinates}

\item{lat}{A numeric vector of latitude coordinates}

\item{x}{A variable to be plotted. If \code{NULL} (the default), the locations only 
will be plotted}

\item{fspdf}{A fortified spatial polygons dataframe}

\item{type}{Either "numeric" or "factor". If \code{NULL} (the default), this function
will try to guess the type of variable. Ignored if both \code{x} and \code{y} are supplied.}

\item{make_simple}{Whether to simplify a categorical variable to fewer than \code{n_simplify}
categories. Ignored unless \code{y} is \code{NULL} and \code{x} is categorical.}

\item{n_simple}{The number of categories to simplify \code{x} to.
Ignored unless \code{y} is \code{NULL} and \code{x} is categorical.}
}
\value{
A plot
}
\description{
Generate simple maps of variables. Meant to show one of 3 different kinds of maps:
1. A simple point map with no coloring
2. A point map with coloring by a categorical variable
3. A point map with coloring by a numeric variable
}

