% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cism_map_interactive.R
\name{cism_map_interactive}
\alias{cism_map_interactive}
\title{interactive cism map}
\usage{
cism_map_interactive(lng, lat, x = NULL, popup = NULL, spdf = NULL,
  type = NULL, make_simple = TRUE, n_simple = 10, opacity = 0.5,
  point_size = 1)
}
\arguments{
\item{lng}{A numeric vector of longitude coordinates}

\item{lat}{A numeric vector of latitude coordinates}

\item{x}{A variable to be plotted. If \code{NULL} (the default), the locations only 
will be plotted}

\item{popup}{A character vector, of identical lenggth to \code{lat} and \code{lng}, 
containing information to be viewed upon clicking}

\item{spdf}{A spatial polygons dataframe}

\item{type}{Either "numeric" or "factor". If \code{NULL} (the default), this function
will try to guess the type of variable. Ignored if both \code{x} and \code{y} are supplied.}

\item{make_simple}{Whether to simplify a categorical variable to fewer than \code{n_simplify}
categories. Ignored unless \code{y} is \code{NULL} and \code{x} is categorical.}

\item{n_simple}{The number of categories to simplify \code{x} to.
Ignored unless \code{y} is \code{NULL} and \code{x} is categorical.}
}
\value{
A plot
}
\description{
Generate simple interactive maps of variables. Meant to show one of 3 different kinds of maps:
1. A simple point map with no coloring
2. A point map with coloring by a categorical variable
3. A point map with coloring by a numeric variable
This function is identical in input to \code{cism_map}, in all ways but three:
1. Rather than returning a ggplot2 plot, it returns a leaflet html widget.
2. It takes a spdf (regular SpatialPolygonsDataFrame) rather than a fspdf (fortified one).
3. It takes an optional popup argument.
}

