% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cism_plot.R
\name{cism_plot}
\alias{cism_plot}
\title{cism plot}
\usage{
cism_plot(x, y = NULL, type = NULL, make_simple = TRUE, n_simple = 10,
  trend = FALSE)
}
\arguments{
\item{x}{A variable to be plotted alone, or the x-axis of a two variable plot}

\item{y}{An option second variable to be plotted on the y-axis}

\item{type}{Either "numeric" or "factor". If \code{NULL} (the default), this function
will try to guess the type of variable. Ignored if both \code{x} and \code{y} are supplied.}

\item{make_simple}{Whether to simplify a categorical variable to fewer than \code{n_simplify}
categories. Ignored unless \code{y} is \code{NULL} and \code{x} is categorical.}

\item{n_simple}{The number of categories to simplify \code{x} to.
Ignored unless \code{y} is \code{NULL} and \code{x} is categorical.}

\item{trend}{Whether to overlay a trend line.
Ignored unless \code{y} is not \code{NULL} and both \code{x} and \code{y} are numeric.}
}
\value{
A plot
}
\description{
Generate simple visualizations of variables. Meant to show one of 3 different kinds of charts:
1. A bar chart (for counts of categorical variables)
2. A histogram (for distributions of numeric variables)
3. A x-y correlation chart (if two variables are supplied)
}
