% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_at_risk.R
\name{create_time_at_risk}
\alias{create_time_at_risk}
\title{Create time at risk}
\usage{
create_time_at_risk(user, password, port = 3306, host = "sap.manhica.net")
}
\arguments{
\item{user}{A username, as a character vector of length one, with access
to the openhds database}

\item{password}{The password corresponding to \code{username}}

\item{port}{The port to use. for the database connection. 
By default 3306; should be 4706 if not at the CISM.}

\item{host}{The host to use for the database connection. 
By default "sap.manhica.net".}
}
\value{
A \code{data.frame} with one row for each person's 
uninterrupted period of residency
}
\description{
Using census migration data, create a dataframe of "time at risk", ie the 
time during which someone lived in Manhiça.
}

