% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_to_bib.R
\name{doi_to_bib}
\alias{doi_to_bib}
\title{Get a bibtex entry from a doi number}
\usage{
doi_to_bib(x = "10.1371/journal.pone.0008500", cat_it = TRUE)
}
\arguments{
\item{x}{A doi}

\item{cat_it}{Whether to cat the output (as opposed to returning a vector)}
}
\value{
A bibtex formatted bibliography entry
}
\description{
Convert a doi (digital object identifier) to a bibtex entry
}

