% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_time_at_risk.R
\name{expand_time_at_risk}
\alias{expand_time_at_risk}
\title{Expand time at risk}
\usage{
expand_time_at_risk(time_at_risk, start_date = "2002-01-01",
  end_date = Sys.Date())
}
\arguments{
\item{time_at_risk}{A dataframe generated by \code{create_time_at_risk}}
}
\value{
A \code{data.frame} with one row for each person's 
uninterrupted period of residency
}
\description{
Using a time at risk dataframe generated by \code{create_time_at_risk},
generate an "expanded" version with one row for each person-day combination
}
\examples{
residency <- get_data(tab = 'residency', dbname = 'openhds')
individual <- get_data(tab = 'individual', dbname = 'openhds')
location <- get_data(tab = 'location', dbname = 'openhds')
time_at_risk <- create_time_at_risk(residency = residency, individual = individual, location = location)
#' expanded_time_at_risk <- expand_time_at_risk(time_at_risk = time_at_risk)
}

