% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Fuzzy match}
\usage{
fuzzy_match(x, y = NULL)
}
\arguments{
\item{x}{A character string}

\item{y}{A character string with which to match
if \code{NULL}, x will be used}
}
\description{
Use approximate string matching to assign "similarity" scores to
potentially mis-entered data
}

