% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionary.R
\name{get_dictionary}
\alias{get_dictionary}
\title{get dictionary}
\usage{
get_dictionary(questions = TRUE, manhica = TRUE, round = 1)
}
\arguments{
\item{questions}{If \code{TRUE}, will return a dictionary of
census questions;
the alternative (\code{FALSE}) is to return the dictionary of the answers}

\item{manhica}{If \code{TRUE}, will return a dictionary for manhica;
the alternative (\code{FALSE}) is to return the dictionary for Magude}

\item{round}{Which round of census (only applicable to Manhica)}
}
\description{
Get a dictionary for the following databases:
1. Manhica (dssodk) with the following tables
HOUSEHOLD_ECONOMICS_CORE,
LOCATION_DETAILS_CORE,
INDIVIDUAL_DETAILS_CORE
2. Magude (round 1) with the following tables
3. Magude (round 2) with the following tables
}
