% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dssodk_dictionary.R
\name{get_dssodk_dictionary}
\alias{get_dssodk_dictionary}
\title{get dssodk dictionary}
\usage{
get_dssodk_dictionary()
}
\description{
Get a dictionary for the following dssodk tables:
HOUSEHOLD_ECONOMICS_CORE, 
LOCATION_DETAILS_CORE,
INDIVIDUAL_DETAILS_CORE
}

