% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maltem_dictionary.R
\name{get_maltem_dictionary}
\alias{get_maltem_dictionary}
\title{get dssodk dictionary}
\usage{
get_maltem_dictionary()
}
\description{
Get a dictionary for the following maltem tables:
HOUSEHOLD
MEMBER
}

