% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_weather}
\alias{get_weather}
\title{Get weather}
\usage{
get_weather(station = "FQMA", start_year = 2014, end_year = 2016,
  save = TRUE, load_saved = TRUE)
}
\arguments{
\item{station}{A weather station. For example, "FQMA" (Maputo), "CDG"
(Charles de Gaulle airport in Paris, France), "JFK" (John F. Kennedy
airport in New York City, USA), etc.}

\item{start_year}{The year for which you would like to start getting data}

\item{end_year}{The year for which you would like to stop getting data}

\item{save}{Whether to save the retrieved data locally}

\item{load_saved}{Whether load, if applicable, previously saved data}
}
\value{
A dataframe of weather data
}
\description{
Get weather data from www.wunderground.com
}
