% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_from_utm.R
\name{ll_from_utm}
\alias{ll_from_utm}
\title{Convert UTM to Latitude Longitude}
\usage{
ll_from_utm(x, y, zone = 36)
}
\arguments{
\item{x}{A numeric vector of x coordinates, or a dataframe with two columns, named
x and y}

\item{y}{A numeric vector of y coordinates (optional; only necessary if x is a vector and not a dataframe)}

\item{zone}{The UTM zone in question. Defaults to 36 (Southern Mozambique)}
}
\value{
A dataframe of latitude and longitude coordinates
}
\description{
Convert from UTM (Universal Transverse Mercator) to Latitude and Longitude
}
