% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mag2}
\alias{mag2}
\title{mag2.}
\format{A SpatialPolygonsDataFrame subset of \code{moz2}}
\usage{
mag2
}
\description{
Map of the Magude district at administrative level 2.
}
\keyword{datasets}
