% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mag2_fortified}
\alias{mag2_fortified}
\title{mag2_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
mag2_fortified
}
\description{
A "fortified" version of \code{mag2}
}
\keyword{datasets}

