% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mag3}
\alias{mag3}
\title{mag3}
\format{A SpatialPolygonsDataFrame subset of \code{moz3}}
\usage{
mag3
}
\description{
Map of the Magude district at administrative level 3.
}
\keyword{datasets}

