% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mag3_fortified}
\alias{mag3_fortified}
\title{mag3_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
mag3_fortified
}
\description{
A "fortified" version of \code{mag3}
}
\keyword{datasets}

