% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{man2}
\alias{man2}
\title{man2.}
\format{A SpatialPolygonsDataFrame subset of \code{moz2}}
\usage{
man2
}
\description{
Map of the Manhiça district at administrative level 2.
}
\keyword{datasets}
