% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{man2_fortified}
\alias{man2_fortified}
\title{man2_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
man2_fortified
}
\description{
A "fortified" version of \code{man2}
}
\keyword{datasets}
