% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{man3}
\alias{man3}
\title{man3.}
\format{A SpatialPolygonsDataFrame subset of \code{moz3}}
\usage{
man3
}
\description{
Map of the Manhiça district at administrative level 3.
}
\keyword{datasets}
