% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{man3_fortified}
\alias{man3_fortified}
\title{man3_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
man3_fortified
}
\description{
A "fortified" version of \code{man3}
}
\keyword{datasets}
