% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mop2}
\alias{mop2}
\title{mop2.}
\format{A SpatialPolygonsDataFrame subset of \code{moz2}}
\usage{
mop2
}
\description{
Map of the Mopeia district at administrative level 2.
}
\keyword{datasets}
