% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mop2_fortified}
\alias{mop2_fortified}
\title{mop2_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
mop2_fortified
}
\description{
A "fortified" version of \code{mop2}
}
\keyword{datasets}

