% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mop3}
\alias{mop3}
\title{mop3}
\format{A SpatialPolygonsDataFrame subset of \code{moz3}}
\usage{
mop3
}
\description{
Map of the Mopeia district at administrative level 3.
}
\keyword{datasets}
