% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mop3_fortified}
\alias{mop3_fortified}
\title{mop3_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
mop3_fortified
}
\description{
A "fortified" version of \code{mop3}
}
\keyword{datasets}

