% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moz0}
\alias{moz0}
\title{moz0.}
\format{A SpatialPolygonsDataFrame with 68 variables in the \code{data} slot : 
OBJECTID
ID_0
ISO
NAME_ENGLISH
NAME_ISO
NAME_FAO
NAME_LOCAL
NAME_OBSOLETE
NAME_VARIANTS
NAME_NONLATIN
NAME_FRENCH
NAME_SPANISH
NAME_RUSSIAN
NAME_ARABIC
NAME_CHINESE
WASPARTOF
CONTAINS
SOVEREIGN
ISO2
WWW
FIPS
ISON
VALIDFR
VALIDTO
POP2000
SQKM
POPSQKM
UNREGION1
UNREGION2
DEVELOPING
CIS
Transition
OECD
WBREGION
WBINCOME
WBDEBT
WBOTHER
CEEAC
CEMAC
CEPLG
COMESA
EAC
ECOWAS
IGAD
IOC
MRU
SACU
UEMOA
UMA
PALOP
PARTA
CACM
EurAsEC
Agadir
SAARC
ASEAN
NAFTA
GCC
CSN
CARICOM
EU
CAN
ACP
Landlocked
AOSIS
SIDS
Islands
LDC

Retrieved from the raster package. For further details on that package, see \url{https://CRAN.R-project.org/package=raster}. Original data source: The GADM database of Global Administrative Areas. For further details on that databse, see \url{http://gadm.org/}.}
\usage{
moz0
}
\description{
Map of Mozambique at administrative level 0.
}
\keyword{datasets}
