% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moz0_fortified}
\alias{moz0_fortified}
\title{moz0_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
moz0_fortified
}
\description{
A "fortified" version of \code{moz0}
}
\keyword{datasets}
