% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moz1}
\alias{moz1}
\title{moz1.}
\format{A SpatialPolygonsDataFrame with the following variables in the \code{data} slot : 
OBJECTID
ID_0
ISO
NAME_0
ID_1
NAME_1
HASC_1
CCN_1
CCA_1
TYPE_1
ENGTYPE_1
NL_NAME_1
VARNAME_1

Retrieved from the raster package. For further details on that package, see \url{https://CRAN.R-project.org/package=raster}. Original data source: The GADM database of Global Administrative Areas. For further details on that databse, see \url{http://gadm.org/}.}
\usage{
moz1
}
\description{
Map of Mozambique at administrative level 1.
}
\keyword{datasets}
