% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moz1_fortified}
\alias{moz1_fortified}
\title{moz1_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
moz1_fortified
}
\description{
A "fortified" version of \code{moz1}
}
\keyword{datasets}

