% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moz2}
\alias{moz2}
\title{moz2.}
\format{A SpatialPolygonsDataFrame with the following variables in the \code{data} slot : 
OBJECTID
ID_0
ISO
NAME_0
ID_1
NAME_1
ID_2
NAME_2
HASC_2
CCN_2
CCA_2
TYPE_2
ENGTYPE_2
NL_NAME_2
VARNAME_2

Retrieved from the raster package. For further details on that package, see \url{https://CRAN.R-project.org/package=raster}. Original data source: The GADM database of Global Administrative Areas. For further details on that databse, see \url{http://gadm.org/}.}
\usage{
moz2
}
\description{
Map of Mozambique at administrative level 2.
}
\keyword{datasets}

