% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moz2_fortified}
\alias{moz2_fortified}
\title{moz2_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
moz2_fortified
}
\description{
A "fortified" version of \code{moz2}
}
\keyword{datasets}
