% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moz3}
\alias{moz3}
\title{moz3.}
\format{A SpatialPolygonsDataFrame with the following variables in the \code{data} slot : 
OBJECTID
ID_0
ISO
NAME_0
ID_1
NAME_1
ID_2
NAME_2
ID_3
NAME_3
CCN_3
CCA_3
TYPE_3
ENGTYPE_3
NL_NAME_3
VARNAME_3
 
Retrieved from the raster package. For further details on that package, see \url{https://CRAN.R-project.org/package=raster}. Original data source: The GADM database of Global Administrative Areas. For further details on that databse, see \url{http://gadm.org/}.}
\usage{
moz3
}
\description{
Map of Mozambique at administrative level 3.
}
\keyword{datasets}
