% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moz3_fortified}
\alias{moz3_fortified}
\title{moz3_fortified}
\format{A dataframe created through \code{ggplot2::fortify}}
\usage{
moz3_fortified
}
\description{
A "fortified" version of \code{moz3}
}
\keyword{datasets}

