% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tables.R
\name{show_tables}
\alias{show_tables}
\title{Show the names of tables of a database}
\usage{
show_tables(dbname = NULL, host = NULL, port = NULL, user = NULL,
  password = NULL, connection_object = NULL, collect = TRUE)
}
\arguments{
\item{dbname}{The name of the database from which you are getting data.
Examples include "openhds", "dss", "dssodk", "maltem_absenteeism", etc. If \code{NULL}
the function will try to use the \code{dbname} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{host}{The name of the host from which you are getting data.
Examples include "sap.manhica.net".
If \code{NULL} the function will try to use the \code{host} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{port}{The name of the port from which you are getting data.
This should be 3306 on-site, and 4706 off-site. If \code{NULL}
the function will try to use the \code{port} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{user}{The user. If \code{NULL}
the function will try to use the \code{user} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{password}{The password If \code{NULL}
the function will try to use the \code{pqssword} in your \code{connection_object}; if the \code{connection_object} is \code{NULL}, the function will
try to create a \code{connection_object} as described below.}

\item{connection_object}{An open connection to a CISM database (as created through \code{credentials_extract} and \code{credentials_connect} or \code{credentials_now}); if \code{NULL}, the function will try to create a \code{connection_object} by retrieving user information from the \code{credentials/credentials.yaml}
in or somewhere upwards of the working directory.}
}
\value{
A dataframe with two columns: \code{dbname} (the database in question)
and \code{table} (the name of the table).
}
\description{
Show the names of tables of a database
}
