% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify a categorical variable}
\usage{
simplify(var, n = 10, other = NA, empty_as_other = FALSE)
}
\arguments{
\item{var}{The factor or character vector to be simplified}

\item{n}{The number of desired levels}

\item{other}{What to replace elements whose counts are not in the top n with}

\item{empty_as_other}{Whether to count empty (ie, "") elements as other}
}
\value{
A factor vector of identical length as var
}
\description{
Simplify the levels of an over-leveled factor by combining small levels.
}

