% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cism.R
\name{theme_cism}
\alias{theme_cism}
\title{CISM theme}
\usage{
theme_cism(base_size = 14, base_family = "Helvetica",
  subtitle_family = "Helvetica", axis_family = "Helvetica")
}
\arguments{
\item{base_size}{The size of the base font}

\item{base_family}{The font used in the legend and title}

\item{subtitle_family}{The font used in the subtitle}

\item{axis_family}{The font used in the axes}
}
\value{
A ggtheme object meant to be used in conjunction with a call to ggplot()
}
\description{
Apply the CISM look to a ggplot2-generated visualization
}

