% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cism_map.R
\name{theme_cism_map}
\alias{theme_cism_map}
\title{CISM theme for maps}
\usage{
theme_cism_map()
}
\value{
A ggtheme object meant to be used in conjunction with a call to ggplot()
}
\description{
Apply the CISM look to a ggplot2-generated visualization. Specific for
maps (no x labels, etc.)
}

