% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_dssodk.R
\name{translate_dssodk}
\alias{translate_dssodk}
\title{translate dssodk}
\usage{
translate_dssodk(dictionary, tab = NULL, table_name, variable)
}
\arguments{
\item{dictionary}{A dictionary dataframe. Should be generated via
\code{\url{get_dssodk_dictionary}}}

\item{tab}{An in-memory dataframe corresponding to the \code{table_name}.
Can be left \code{NULL} if \code{table_name} is identical to the name of the object in memory.}

\item{table_name}{The name (as a character vector of length one) of the table
to be translated}

\item{variable}{The name (as a character vector of length one) of the
variable to be translated}
}
\value{
A one column dataframe in which the variable name will be the
Portugese language question pertaining to the \code{variable}, and
the values of the dataframe will be the Portuguese-language values in the
dictionary
}
\description{
Translate columns of dssodk.
}
